cloud_functions/events/requirements.txt
functions-framework
google-cloud-pubsub
google-cloud-logging
google-cloud-secret-manager
requests==2.25.1
pytz==2021.1
python-dateutil==2.8.2
jsonschema==4.23.0
sqlalchemy
cloud-sql-python-connector
pg8000
python-slugify
beautifulsoup4==4.13.3